+++ src/bindings/python/src/predicted_speeds_module.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <nanobind/nanobind.h>
+
+namespace pyvalhalla {
+void init_predicted_speeds(nanobind::module_& m);
+}